Apply pr request feedback
