+++ core/src/syntax.rs
@@ -15,6 +15,7 @@
 // along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
 use async_smtp::EmailAddress;
+use levenshtein::levenshtein;
 use serde::{Deserialize, Serialize};
 use std::str::FromStr;
 
@@ -95,6 +96,33 @@ pub fn check_syntax(email_address: &str) -> SyntaxDetails {
 	}
 }
 
+const MAIL_PROVIDERS: &[&str] = &[
+	"gmail.com",
+	"yahoo.com",
+	"outlook.com",
+	"hotmail.com",
+	"protonmail.com",
+	"icloud.com",
+	"yandex.com",
+];
+// Supplies the syntax parameter with a suggestion that matches the mail domain best by levenshtein
+// distance.
+pub fn get_similar_mail_provider(syntax: &mut SyntaxDetails) {
+	for possible_provider in MAIL_PROVIDERS {
+		let distance = levenshtein(&syntax.domain, possible_provider);
+
+		if distance < 3 {
+			// Return full address
+			syntax.suggestion = Some(format!(
+				"{}@{}",
+				syntax.username,
+				String::from(*possible_provider),
+			));
+			break;
+		}
+	}
+}
+
 #[cfg(test)]
 mod tests {
 	use super::*;
@@ -139,5 +167,18 @@ mod tests {
 				suggestion: None,
 			}
 		);
+	}
+
+	#[test]
+	fn should_suggest_a_correct_mail_if_similar() {
+		let mut syntax = SyntaxDetails {
+			address: Some(EmailAddress::new("test@gmali.com".into()).unwrap()),
+			domain: "gmali.com".into(),
+			is_valid_syntax: true,
+			username: "test".into(),
+			suggestion: None,
+		};
+		get_similar_mail_provider(&mut syntax);
+		assert_eq!(syntax.suggestion, Some("test@gmail.com".to_string()))
 	}
 }
